+++ server/components/user/__test__/user-routes.spec.js
@@ -3,6 +3,7 @@
  * Test for user component
  * @module user-routes.spec
  */
+
 'use strict'
 
 // Set the env variable to test and load server
@@ -11,7 +12,7 @@ const server = require('../../../server')
 
 const chai = require('chai')
 const chaiHttp = require('chai-http')
-const checkRequestError = require('../../../helpers/test-api').checkRequestError
+const checkRequestError = require('../../../../test/helpers/test-api').checkRequestError
 const userDAO = require('../user.dao')
 
 const app = server.app
@@ -23,7 +24,7 @@ const expect = chai.expect
 // Parent block
 describe('User routes', () => {
   // Remove all related with users
-  before(done => {
+  before((done) => {
     userDAO
       .removingAllUsers()
       .then(() => done())
@@ -36,7 +37,7 @@ describe('User routes', () => {
     const url = '/api/users'
 
     // Error responses
-    it('it should not POST a user without any data', done => {
+    it('it should not POST a user without any data', (done) => {
       const user = {}
       const expectedFields = ['email', 'role']
       const request = chai.request(app).post(url).send(user)
@@ -44,7 +45,7 @@ describe('User routes', () => {
     })
 
     // Create a user
-    it('it should POST a user', done => {
+    it('it should POST a user', (done) => {
       const user = {
         email: 'user@mail.test',
         role: 'user'
@@ -61,7 +62,7 @@ describe('User routes', () => {
           done()
         })
     })
-    it('it should not POST a user if already exits', done => {
+    it('it should not POST a user if already exits', (done) => {
       const user = {
         email: 'user@mail.test',
         role: 'user'
@@ -70,7 +71,7 @@ describe('User routes', () => {
       const request = chai.request(app).post(url).send(user)
       checkRequestError(request, 422, 'ValidationError', expectedFields, done)
     })
-    it('it should POST a second admin', done => {
+    it('it should POST a second admin', (done) => {
       const user = {
         email: 'admin@mail.test',
         role: 'admin'
@@ -96,7 +97,7 @@ describe('User routes', () => {
     const url = '/api/users'
 
     // Get all users
-    it('it should GET all users', done => {
+    it('it should GET all users', (done) => {
       chai.request(app)
         .get(url)
         .end((err, res) => {
@@ -113,18 +114,18 @@ describe('User routes', () => {
    */
   describe('GET /api/users/:userID', () => {
     const url = '/api/users'
-    it('it should not GET a user if not find id', done => {
+    it('it should not GET a user if not find id', (done) => {
       const request = chai.request(app).get(`${url}/000000000000000000000000`)
-      checkRequestError(request, 404, 'UserNotFound', done)
+      checkRequestError(request, 404, 'UserNotFound', null, done)
     })
-    it('it should GET a user by the given id', done => {
+    it('it should GET a user by the given id', (done) => {
       const user = {
         email: 'user_to_getbyid@mail.test',
         role: 'user'
       }
       userDAO
         .creatingUser(user)
-        .then(newUser => {
+        .then((newUser) => {
           chai.request(app)
             .get(`${url}/${newUser._id}`)
             .end((err, res) => {