+++ server/components/user/user-error.map.js
@@ -2,6 +2,7 @@
  * A map with custom error for user component
  * @module user-error.map
  */
+
 'use strict'
 
 // Create custom errors