+++ server/config/config.js
@@ -1,6 +1,7 @@
 /**
  * @module config
  */
+
 'use strict'
 
 /**