+++ server/sandbox/sandbox.spec.js
@@ -3,6 +3,7 @@
  * Test for sandbox module
  * @module sandbox.spec
  */
+
 'use strict'
 
 // Set the env variable to test